Делит код на модули и настраивает сборку с помощью webpack
