+++ public/index.html
@@ -1,20 +1,18 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
+
+<head>
     <meta charset="UTF-8" />
-    <meta
-      name="viewport"
-      content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0"
-    />
+    <meta name="viewport" content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0" />
     <meta http-equiv="X-UA-Compatible" content="ie=edge" />
     <link rel="stylesheet" href="css/normalize.css" />
     <link rel="stylesheet" href="css/style.css" />
     <title>TASKMANAGER</title>
-  </head>
+</head>
 
-  <body>
+<body>
     <div style="display: none">
-      <svg>
+        <svg>
         <symbol
           id="wave"
           viewBox="0 0 159 10"
@@ -35,15 +33,16 @@
     </div>
 
     <main class="main">
-      <section class="main__control control container">
-        <h1 class="control__title">TASKMANAGER</h1>
-        <!-- Меню -->
-      </section>
+        <section class="main__control control container">
+            <h1 class="control__title">TASKMANAGER</h1>
+            <!-- Меню -->
+        </section>
 
-      <!-- Фильтры -->
+        <!-- Фильтры -->
 
-      <!-- Контент -->
+        <!-- Контент -->
     </main>
-    <script src="../src/main.js"></script>
-  </body>
+    <script src="./bundle.js"></script>
+</body>
+
 </html>
