+++ src/components/site-button-load.js
@@ -1,0 +1,7 @@
+const createButtonLoad = () => {
+  return (
+    `<button class="load-more" type="button">load more</button>`
+  );
+};
+
+export {createButtonLoad};
